fix: show automation activity on issue page
